+++ addon/components/bsy-nav.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+import layout from '../templates/components/bsy-nav';
+
+export default Ember.Component.extend({
+  layout,
+  tagName: 'ul',
+  classNames: 'nav'
+});